# Reading/Writing GeoJSON via GEOSCoordSequence_copy(From|To)Buffer (#17)