--- a/docs/.vitepress/cache/deps/@theme_index.js
+++ b/docs/.vitepress/cache/deps/@theme_index.js
@@ -13,27 +13,27 @@ import {
 import "./chunk-76J2PTFD.js";
 
 // node_modules/vitepress/dist/client/theme-default/index.js
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/fonts.css";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/fonts.css";
 
 // node_modules/vitepress/dist/client/theme-default/without-fonts.js
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/vars.css";
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/base.css";
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/utils.css";
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/custom-block.css";
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/vp-code.css";
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/vp-code-group.css";
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/vp-doc.css";
-import "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/vp-sponsor.css";
-import VPBadge from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPBadge.vue";
-import Layout from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/Layout.vue";
-import { default as default2 } from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPHomeHero.vue";
-import { default as default3 } from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPHomeFeatures.vue";
-import { default as default4 } from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPHomeSponsors.vue";
-import { default as default5 } from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPDocAsideSponsors.vue";
-import { default as default6 } from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPTeamPage.vue";
-import { default as default7 } from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPTeamPageTitle.vue";
-import { default as default8 } from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPTeamPageSection.vue";
-import { default as default9 } from "/Users/pahmeyer/Documents/GitHub/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPTeamMembers.vue";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/vars.css";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/base.css";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/utils.css";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/custom-block.css";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/vp-code.css";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/vp-code-group.css";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/vp-doc.css";
+import "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/styles/components/vp-sponsor.css";
+import VPBadge from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPBadge.vue";
+import Layout from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/Layout.vue";
+import { default as default2 } from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPHomeHero.vue";
+import { default as default3 } from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPHomeFeatures.vue";
+import { default as default4 } from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPHomeSponsors.vue";
+import { default as default5 } from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPDocAsideSponsors.vue";
+import { default as default6 } from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPTeamPage.vue";
+import { default as default7 } from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPTeamPageTitle.vue";
+import { default as default8 } from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPTeamPageSection.vue";
+import { default as default9 } from "/Users/pahmeyer/Documents/GitHub.nosync/geos-wasm/node_modules/vitepress/dist/client/theme-default/components/VPTeamMembers.vue";
 
 // node_modules/@vueuse/shared/index.mjs
 function tryOnScopeDispose(fn) {