Hi @chrispahm,
I implemented #17 on this PR branch, and I confirmed that `npm run performance` result was a bit (1.5x) faster than `wkb(wkx)` one, so just for your information. 😄

| Test Pattern                      | Current wkb(wkx) | Use this PR | @turf/turf (2 times) |
|--------------------------|-------------------|-------------|--------------------|
| Testing feature Simple     | 29 ms |  **18 ms** |  54/56 ms |
| Testing feature Australia  | 5204 ms | **3098 ms** | 6473/6465 ms |
| Real life performance test | 300 ms | **188 ms** | 244/238 ms |
| Buffering each geometry individually | 258 ms | **134 ms** | 141/144 ms |

`Use this PR` needs the following changes in `test/performance.mjs`.
```diff
--- a/test/performance.mjs
+++ b/test/performance.mjs
@@ -2,7 +2,7 @@ import { performance } from 'perf_hooks'
 import * as turf from '@turf/turf'
 import { readFileSync } from 'fs'
 import initGeosJs from '../src/index.mjs'
-import geosBuffer from '../docs/assets/buffer.js'
+import geosBuffer from '../docs/assets/buffer2.js'
```

<details>

### Current wkb(wkx)

```sh
% npm run performance

> geos-wasm@1.1.4 performance
> node test/performance.mjs

Testing feature Simple
GEOS buffer method took 29 milliseconds to run 100 times.
Average time per iteration: 0 milliseconds.
@turf/turf buffer method took 54 milliseconds to run 100 times.
Average time per iteration: 1 milliseconds.
GEOS buffer method was 1.86x faster than @turf/turf buffer method.

Testing feature Australia
GEOS buffer method took 5204 milliseconds to run 100 times.
Average time per iteration: 52 milliseconds.
@turf/turf buffer method took 6473 milliseconds to run 100 times.
Average time per iteration: 65 milliseconds.
GEOS buffer method was 1.24x faster than @turf/turf buffer method.

Real life performance test
Buffering all geometries at once
GEOS buffer method took 300 milliseconds to run 1 times.
Average time per iteration: 300 milliseconds.
@turf/turf buffer method took 244 milliseconds to run 1 times.
Average time per iteration: 244 milliseconds.
@turf/turf buffer method was 1.23x faster than GEOS buffer method.

Buffering each geometry individually
GEOS buffer method took 258 milliseconds to run 1 times.
Average time per iteration: 258 milliseconds.
@turf/turf buffer method took 141 milliseconds to run 1 times.
Average time per iteration: 141 milliseconds.
@turf/turf buffer method was 1.83x faster than GEOS buffer method.
```

### Use this PR version

```sh
% npm run performance

> geos-wasm@1.1.4 performance
> node test/performance.mjs

Testing feature Simple
GEOS buffer method took 18 milliseconds to run 100 times.
Average time per iteration: 0 milliseconds.
@turf/turf buffer method took 56 milliseconds to run 100 times.
Average time per iteration: 1 milliseconds.
GEOS buffer method was 3.11x faster than @turf/turf buffer method.

Testing feature Australia
GEOS buffer method took 3098 milliseconds to run 100 times.
Average time per iteration: 31 milliseconds.
@turf/turf buffer method took 6465 milliseconds to run 100 times.
Average time per iteration: 65 milliseconds.
GEOS buffer method was 2.09x faster than @turf/turf buffer method.

Real life performance test
Buffering all geometries at once
GEOS buffer method took 188 milliseconds to run 1 times.
Average time per iteration: 188 milliseconds.
@turf/turf buffer method took 238 milliseconds to run 1 times.
Average time per iteration: 238 milliseconds.
GEOS buffer method was 1.27x faster than @turf/turf buffer method.

Buffering each geometry individually
GEOS buffer method took 134 milliseconds to run 1 times.
Average time per iteration: 134 milliseconds.
@turf/turf buffer method took 144 milliseconds to run 1 times.
Average time per iteration: 144 milliseconds.
GEOS buffer method was 1.07x faster than @turf/turf buffer method.
```

</details>

Before the implementation, I wrote some Coordinate Sequence and Geometry Constructor tests to be familiar with the usage and I fixed/adjusted some parts, but those can be separated to another PR, so if you feel that this changes are too much, let me know about that.