@sanak 2023-08-08
@chrispahm Thanks for confirmation!

> Thanks again for this awesome PR! There are a few tiny changes I'd ask for, as outlined in the review comments. Other things that would be nice would be to
> 
> * incorporate this method into the docs examples
> * Update the FAQ https://chrispahm.github.io/geos-wasm/#what-is-the-preferred-way-to-pass-geometries-between-javascript-and-geos-wasm to reflect this more performant option
> * ?
> 
> Let me know if you want me to take over any of those tasks!

Thanks!
Because I am still not familiar with docs/examples and build part, taking over above tasks is really helpful for me. 🙇🙏
I will reply the code comments, so please wait for a while.