@sanak 2023-08-09
@chrispahm Well, I will take one week summer holidays from tomorrow, so I am okay to do remaining things, if you haven't started update yet.
Though my English writing and understanding docs/examples and build part may take a certain time. 😊