@sanak 2023-08-11
@chrispahm
I am still on about halfway, but I get stuck...

I renamed all files/methods as you mentioned, so those part should be solved, but when I switched from `buffer_wkb.js` to `buffer_coordseq.js` in `docs/components/buffer.vue`, I got the following browser console error from `npm run docs:dev`.
```
geojson.mjs:322 Uncaught TypeError: geos.Module._malloc is not a function
    at geojsonCoordsToGeosCoordSeq (geojson.mjs:322:32)
    at geojsonToGeosGeom (geojson.mjs:65:26)
    at bufferFeature (buffer_coordseq.js:204:19)
    at buffer_coordseq.js:107:31
    at featureEach (index.js:395:11)
    at buffer (buffer_coordseq.js:106:7)
    at updateBuffer (buffer.vue:256:26)
    at Ba.Map.<anonymous> (buffer.vue:296:5)
    at Ba.Map.fire (evented.ts:130:26)
    at Ba.Map._render (map.ts:3194:18)
```

In original `buffer_wkb.js`, `geos` object passed to `GEOSFunctions` via `buffer` method's option property (`GEOS`), but I don't know the best way to pass the object to new helper...

Probably, the easiest way will be adding `geos` object argument at each helper methods (the following diff image),
```diff
--- a/src/helpers/geojson.mjs
+++ b/src/helpers/geojson.mjs
@@ -1,6 +1,4 @@
-import { geos } from '../allCFunctions.mjs'
-
-export function geojsonToGeosGeom (geojson) {
+export function geojsonToGeosGeom (geos, geojson) {
   // assume only 2d (x, y) geometries
   switch (geojson.type) {
     case 'Feature':
@@ -165,7 +163,7 @@ export function geojsonToGeosGeom (geojson) {
   }
 }
 
-export function geosGeomToGeojson (geomPtr) {
+export function geosGeomToGeojson (geos, geomPtr) {
   if (!geomPtr) {
     return null
   }
@@ -312,7 +310,7 @@ export function geosGeomToGeojson (geomPtr) {
   }
 }
 
-function geojsonCoordsToGeosCoordSeq (coords) {
+function geojsonCoordsToGeosCoordSeq (geos, coords) {
   if (!coords || coords.length === 0) {
     return null
   }
@@ -326,7 +324,7 @@ function geojsonCoordsToGeosCoordSeq (coords) {
   return seqPtr
 }
 
-function geosCoordSeqToGeojsonCoords (seqPtr) {
+function geosCoordSeqToGeojsonCoords (geos, seqPtr) {
   if (!seqPtr) {
     return null
   }
```
then pass `geos` argument from client app side, but how do you think about this ?