@sanak 2023-08-12
@chrispahm About the above `buffer_coordseq.js` issue in `docs/components/buffer.vue`, other solution will be exporting 2 methods via `geos.helpers` object (or namespace).
I prefer this way than above adding `geos` argument` way, but let me know your preference.
```diff
--- a/src/index.mjs
+++ b/src/index.mjs
@@ -1,4 +1,5 @@
 import { initCFunctions, geos } from './allCFunctions.mjs'
+import { geojsonToGeosGeom, geosGeomToGeojson } from './helpers/geojson.mjs'
 import CModule from '../build/package/geos.js'
 
 let geosJsPromise
@@ -30,6 +31,10 @@ export default function initGeosJs (config = {}) {
     }
 
     CModule(geos.Module).then((res) => {
+      geos.helpers = {
+        geojsonToGeosGeom,
+        geosGeomToGeojson
+      }
       resolve(geos)
     })
   })
```
```diff
--- a/docs/assets/buffer_coordseq.js
+++ b/docs/assets/buffer_coordseq.js
@@ -33,7 +33,6 @@ import {
   lengthToRadians,
   earthRadius
 } from '@turf/helpers'
-import { geojsonToGeosGeom, geosGeomToGeojson } from '../../src/helpers/geojson.mjs'
 let GEOSFunctions
 
 /**
@@ -201,7 +200,7 @@ function bufferFeature (geojson, radius, units, steps, endCapStyle, joinStyle, m
     }
   }
   // create a GEOS object from the GeoJSON
-  const geomPtr = geojsonToGeosGeom(projected)
+  const geomPtr = GEOSFunctions.helpers.geojsonToGeosGeom(projected)
   const distance = radiansToLength(lengthToRadians(radius, units), 'meters')
   let bufferPtr
   if (isBufferWithParams) {
@@ -214,7 +213,7 @@ function bufferFeature (geojson, radius, units, steps, endCapStyle, joinStyle, m
     GEOSFunctions.GEOSBufferParams_destroy(bufferParamsPtr)
   }
   // update the original GeoJSON with the new geometry
-  const buffered = geosGeomToGeojson(bufferPtr)
+  const buffered = GEOSFunctions.helpers.geosGeomToGeojson(bufferPtr)
   // destroy the GEOS objects
   GEOSFunctions.GEOSGeom_destroy(geomPtr)
   GEOSFunctions.GEOSGeom_destroy(bufferPtr)
```